The displayed window was implemented a function was created to display the window when someone push the button all elements of the js files were merged in one file were deteled the empty files and folders comments were deleted from html files
