+++ index.html
@@ -454,5 +454,5 @@
       </footer>
     </div>
   </body>
-  <script src="./js/script.js" type="module"></script>
+  <script src="./script.js" type="module"></script>
 </html>
