--- js/modules/allData.js
@@ -1,111 +1,0 @@
-const allData = [
-  {
-    data1: {
-      title: 'My Recent Works',
-      subtitle: 'Multi-Post Stories',
-      tools: {
-        0: 'css',
-        1: 'html',
-        2: 'bootstrap',
-        3: 'Ruby',
-      },
-      image: './Images/Snapshoot Portfolio.png',
-      description: "A daily selection of privately personalized reads; no accounts or sign-ups required. has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a standard dummy text.",
-      source: 'https://www.google.com/',
-      live: 'https://www.facebook.com/',
-    },
-  },
-  {
-    data2: {
-      title: 'Profesional Art Printing Data',
-      tools: {
-        0: 'html',
-        1: 'Bootstrap',
-        2: 'Ruby',
-      },
-      image: './Images/project1.jpeg',
-      description:
-        "A daily selection of privately personalized reads; no accounts or sign-ups required. has been the industry's standard",
-      source: '#',
-      live: '#',
-    },
-  },
-  {
-    data3: {
-      title: 'Data Dashboard<br>Healthcare',
-      mobileTitle: 'Profesional Art Printing Data',
-      tools: {
-        0: 'html',
-        1: 'Bootstrap',
-        2: 'Ruby',
-      },
-      image: './Images/project2.png',
-      description: "A daily selection of privately personalized reads; no accounts or sign-ups required. has been the industry's standard",
-      source: '#',
-      live: '#',
-    },
-  },
-  {
-    data4: {
-      title: 'Website Protfolio ',
-      mobileTitle: 'Profesional Art Printing Data',
-      tools: {
-        0: 'html',
-        1: 'Bootstrap',
-        2: 'Ruby',
-      },
-      image: './Images/project3.png',
-      description: "A daily selection of privately personalized reads; no accounts or sign-ups required. has been the industry's standard",
-      source: '#',
-      live: '#',
-    },
-  },
-  {
-    data5: {
-      title: 'Profesional Art<br>Printing Data More',
-      mobileTitle: 'Profesional Art Printing Data',
-      tools: {
-        0: 'html',
-        1: 'Bootstrap',
-        2: 'Ruby',
-      },
-      image: './Images/project4.jpg',
-      description: "A daily selection of privately personalized reads; no accounts or sign-ups required. has been the industry's standard",
-      source: '#',
-      live: '#',
-    },
-  },
-  {
-    data6: {
-      title: 'Data Dashboard<br>Healthcare',
-      mobileTitle: 'Profesional Art Printing Data',
-      tools: {
-        0: 'html',
-        1: 'Bootstrap',
-        2: 'Ruby',
-      },
-      image: './Images/project5.png',
-      description: "A daily selection of privately personalized reads; no accounts or sign-ups required. has been the industry's standard",
-      source: '#',
-      live: '#',
-    },
-  },
-  {
-    data7: {
-      title: 'Website Protfolio',
-      mobileTitle: 'Profesional Art Printing Data',
-      tools: {
-        0: 'html',
-        1: 'Bootstrap',
-        2: 'Ruby',
-      },
-      image: './Images/project6.jpeg',
-      description:
-        "A daily selection of privately personalized reads; no accounts or sign-ups required. has been the industry's standard",
-      source: '#',
-      live: '#',
-    },
-  },
-];
-
-export { allData };
